cache moved to common, added for meta analyzer
