+++ repository-meta-analyzer/src/main/java/org/acme/repositories/GemMetaAnalyzer.java
@@ -23,6 +23,7 @@ import com.github.packageurl.PackageURL;
 import kong.unirest.*;
 import org.acme.common.UnirestFactory;
 import org.acme.model.Component;
+import org.acme.model.MetaModel;
 import org.acme.model.RepositoryType;
 
 import javax.enterprise.context.ApplicationScoped;
