+++ repository-meta-analyzer/src/main/java/org/acme/repositories/HexMetaAnalyzer.java
@@ -25,6 +25,7 @@ import kong.unirest.json.JSONArray;
 import kong.unirest.json.JSONObject;
 import org.acme.common.UnirestFactory;
 import org.acme.model.Component;
+import org.acme.model.MetaModel;
 import org.acme.model.RepositoryType;
 
 import javax.enterprise.context.ApplicationScoped;
