+++ repository-meta-analyzer/src/main/java/org/acme/repositories/MavenMetaAnalyzer.java
@@ -20,6 +20,7 @@ package org.acme.repositories;
 
 import alpine.common.logging.Logger;
 import com.github.packageurl.PackageURL;
+import org.acme.model.MetaModel;
 import org.apache.http.HttpEntity;
 import org.apache.http.StatusLine;
 import org.apache.http.client.methods.CloseableHttpResponse;
