+++ repository-meta-analyzer/src/main/java/org/acme/repositories/NugetMetaAnalyzer.java
@@ -23,6 +23,7 @@ import com.github.packageurl.PackageURL;
 import kong.unirest.*;
 import kong.unirest.json.JSONArray;
 import kong.unirest.json.JSONObject;
+import org.acme.model.MetaModel;
 import org.apache.maven.artifact.versioning.ComparableVersion;
 import org.acme.common.UnirestFactory;
 import org.acme.model.Component;
